requirements.txt
streamlit==1.19.0
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
altair==4.2.2